Moves old lxdclient/utils functionality to container/lxd. Recruits this to check if LXD is running when updating proxies. Includes a few drive-by spelling fixes.
